+++ test/support/test_case.ex
@@ -1,4 +1,5 @@
 defmodule Timber.TestCase do
+  @moduledoc false
   use ExUnit.CaseTemplate, async: false
 
   setup _tags do